tsdb: type checking per series
This commit implements doing type checking per series. Since we store
the type information in the series id now, we can use that to ensure
that it doesn't change.

A large portion of the refactoring is to have the code use a new
abstraction: the tsdb.SeriesCollection. This makes it so that we can
avoid passing around 4 parallel slices.

The other major portion of the complexity comes from how the series
file handles doing lookups/writes of new series in parallel. Since
that's also the spot where we know what type the series id is supposed
to be, invalidating the series becomes challenging. Using the fact
that most writes won't have a problem, we make it cheap and allocation
free in the case that there are no conflicts.

The last change that was made was to always use the value form of
a tsdb.PartialWriteError, as that is what upper layers were expecting.

Some benchmark results for inmem[1], names elided for space:

```
old time/op    new time/op    delta
  2.70ms ± 1%    2.64ms ± 1%  -2.31%          (p=0.000 n=8+7)
   314µs ± 2%     315µs ± 1%    ~             (p=0.442 n=8+8)
   314µs ± 4%     315µs ± 1%    ~             (p=0.536 n=8+7)
  2.02ms ± 1%    2.00ms ± 1%  -0.88%          (p=0.001 n=8+7)

old alloc/op   new alloc/op   delta
   477kB ± 0%     477kB ± 0%  -0.04%          (p=0.000 n=8+8)
   15.0B ± 0%     15.0B ± 0%    ~     (all samples are equal)
   15.0B ± 0%     15.0B ± 0%    ~     (all samples are equal)
   501kB ± 0%     501kB ± 0%  -0.04%          (p=0.000 n=8+8)

old allocs/op  new allocs/op  delta
    37.0 ± 0%      38.0 ± 0%  +2.70%          (p=0.000 n=8+8)
   0.00 ±NaN%     0.00 ±NaN%    ~     (all samples are equal)
   0.00 ±NaN%     0.00 ±NaN%    ~     (all samples are equal)
    23.0 ± 0%      23.0 ± 0%    ~     (all samples are equal)
```

tsi1 has a more comprehensive benchmark suite, but I had trouble
running it, so I used inch to create 30 million series by running
`inch -b 10000 -c 20 -t 10,10,100000,3 -m 1 -p 1 -v`:

| strategy    | vals/sec          | seconds        | ratio |
|-------------|-------------------|----------------|-------|
| no checking | 80829.1 val/sec.  | 371.2 seconds  | 1x    |
| checking    | 81896.5 val/sec.  | 366.3 seconds  | 0.98x |

which I believe can be attributed to noise.

[1] The inmem bennchmarks were also benchmarking the initial add in
some situations, causing the allocation amounts to vary based on how
many times it ran, since only the first time allocated. The above
benchmarks include backported changes to fix this.